WIP: Add KMS IT test.
